requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
idna==2.7
psycopg2-binary==2.7.5
requests==2.19.1
selenium==3.14.1
urllib3==1.23
